+++ tests/conftest.py
@@ -159,6 +159,7 @@ def cyclopentane_cyclohexane():
 def propane_cyclopropane():
     return sr.load_test_files("propane_cyclopropane.bss")
 
+
 @pytest.fixture(scope="session")
 def pentane_cyclopentane():
     return sr.load_test_files("pentane_cyclopentane.bss")